backend/requirements.txt
asgiref==3.8.1
celery==5.3.6
Django==5.1.3
django-celery-beat
django-configurations==2.5.1
django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.15.2
django-filter==25.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flake8==7.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.2
redis
